+++ circuits/CHANGELOG.md
@@ -17,6 +17,7 @@ verification keys break backwards compatibility.
 * Change nr of bits to represent JubJub scalar field modulus from 255 -> 252 [#179](https://github.com/midnightntwrk/midnight-zk/pull/179)
 
 ### Removed
+* Move external implementations to zk-stdlib [#178](https://github.com/midnightntwrk/midnight-zk/pull/178) 
 
 ## [6.0.0] - 18-12-2025
 ### Added